sms/sms-marketing/requirements.txt
commune-mail>=0.3.0
openai>=1.30.0
python-dotenv>=1.0.0
